refactor: Step 1.1 - Extract PostCard component from HomeScreen (reduces HomeScreen by ~150 lines)
