+++ screens/HomeScreen.js
@@ -22,6 +22,7 @@ import { userService, postService } from '../services/api';
 import { workoutAPI } from '../services/workoutAPI';
 import { useTheme } from '../context/ThemeContext';
 import { useFocusEffect } from '@react-navigation/native';
+import PostCard from '../components/social/PostCard';
 
 // Replace with your actual backend URL
 const API_URL = __DEV__ 
@@ -29,210 +30,6 @@ const API_URL = __DEV__
   : 'https://trainly-backend-production.up.railway.app';  // Production
 const USER_DATA_KEY = '@user_data';
 
-// Helper function to safely handle image URIs of various formats
-const getSafeImageUri = (imageSource) => {
-  // If it's already a require statement (local image), return as is
-  if (typeof imageSource !== 'string') {
-    return imageSource;
-  }
-
-  // Handle file:// URIs (from device storage)
-  if (imageSource && imageSource.startsWith('file://')) {
-    return { uri: imageSource };
-  }
-  
-  // Handle paths that start with "/data/" (internal storage paths)
-  if (imageSource && imageSource.startsWith('/data/')) {
-    return { uri: `file://${imageSource}` };
-  }
-  
-  // Handle paths that directly reference ExperienceData
-  if (imageSource && imageSource.includes('ExperienceData')) {
-    return { uri: imageSource };
-  }
-  
-  // Check if the URL starts with http or https
-  if (imageSource && (imageSource.startsWith('http://') || imageSource.startsWith('https://'))) {
-    return { uri: imageSource };
-  }
-  
-  // If it's a local path without http, add the base URL
-  if (imageSource && !imageSource.startsWith('/')) {
-    return { uri: `${API_URL}/${imageSource}` };
-  }
-  
-  // If it starts with /, assume it's a local path on the server
-  if (imageSource && imageSource.startsWith('/')) {
-    return { uri: `${API_URL}${imageSource}` };
-  }
-  
-  // Fallback to default image
-  return require('../assets/images/bike.jpg');
-};
-
-// Helper function to safely extract the user ID for navigation
-const getUserIdFromPost = (post) => {
-  if (!post || !post.user) return null;
-  
-  // Check for different ID field formats
-  if (post.user._id) return post.user._id;
-  if (post.user.id) return post.user.id;
-  if (post.user.userId) return post.user.userId;
-  
-  // If the user object itself is an ID string
-  if (typeof post.user === 'string') return post.user;
-  
-  // If the post has a userId field
-  if (post.userId) return post.userId;
-  
-  return null;
-};
-
-// PostCard component for your HomeScreen
-const PostCard = ({ post, onLike, navigation }) => {
-  const theme = useTheme();
-  const colors = theme.colors;
-
-  const handleLike = async () => {
-    if (onLike) {
-      onLike(post.id || post._id);
-    }
-  };
-
-  const handleComment = () => {
-    navigation.navigate('Comments', {
-      postId: post.id || post._id,
-      postContent: post.content,
-      postUser: post.user
-    });
-  };
-
-  // Navigate to user profile when avatar or name is clicked
-  const handleUserPress = () => {
-    const userId = getUserIdFromPost(post);
-    
-    if (userId) {
-      navigation.navigate('UserProfile', { userId });
-    } else {
-      console.log('Cannot navigate to profile - missing user ID');
-      console.log('Post user structure:', JSON.stringify(post.user, null, 2));
-    }
-  };
-
-  // Format the date for display
-  const formatDate = (dateString) => {
-    const date = new Date(dateString);
-    const now = new Date();
-    const diffTime = Math.abs(now - date);
-    const diffDays = Math.floor(diffTime / (1000 * 60 * 60 * 24));
-    const diffHours = Math.floor(diffTime / (1000 * 60 * 60));
-    const diffMinutes = Math.floor(diffTime / (1000 * 60));
-    
-    if (diffDays > 0) {
-      return `${diffDays} day${diffDays > 1 ? 's' : ''} ago`;
-    } else if (diffHours > 0) {
-      return `${diffHours} hour${diffHours > 1 ? 's' : ''} ago`;
-    } else {
-      return `${diffMinutes} minute${diffMinutes > 1 ? 's' : ''} ago`;
-    }
-  };
-
-  // Get user display name safely
-  const getUserName = () => {
-    if (!post.user) return 'User';
-    
-    if (typeof post.user === 'string') return 'User';
-    
-    return post.user.name || post.user.username || 'User';
-  };
-
-  return (
-    <View style={[styles.postCard, { backgroundColor: colors.surface }]}>
-      <View style={styles.postHeader}>
-        <TouchableOpacity onPress={handleUserPress}>
-          <Image 
-            source={post.user && post.user.avatar 
-              ? getSafeImageUri(post.user.avatar) 
-              : require('../assets/images/bike.jpg')} 
-            style={styles.userAvatar} 
-            onError={(e) => {
-              console.log('Avatar image error:', e.nativeEvent.error);
-            }}
-          />
-        </TouchableOpacity>
-        <View style={styles.userInfo}>
-          <TouchableOpacity onPress={handleUserPress}>
-            <Text style={[styles.userName, { color: colors.text }]}>{getUserName()}</Text>
-          </TouchableOpacity>
-          <Text style={[styles.postTime, { color: colors.textSecondary }]}>
-            {post.createdAt ? formatDate(post.createdAt) : '2 hours ago'}
-          </Text>
-        </View>
-      </View>
-      
-      {post.content && <Text style={[styles.postContent, { color: colors.text }]}>{post.content}</Text>}
-      
-      {/* Workout details section (if available) */}
-      {post.workoutDetails && post.workoutDetails.type && (
-        <View style={[styles.workoutDetailsContainer, { backgroundColor: `${colors.secondary}15` }]}>
-          <View style={styles.workoutTypeContainer}>
-            <Ionicons name="fitness-outline" size={16} color={colors.primary} />
-            <Text style={[styles.workoutType, { color: colors.primary }]}>{post.workoutDetails.type}</Text>
-          </View>
-          
-          <View style={styles.workoutStatsContainer}>
-            {post.workoutDetails.duration > 0 && (
-              <View style={styles.workoutStat}>
-                <Ionicons name="time-outline" size={14} color={colors.textSecondary} />
-                <Text style={[styles.workoutStatText, { color: colors.textSecondary }]}>{post.workoutDetails.duration} min</Text>
-              </View>
-            )}
-            
-            {post.workoutDetails.calories > 0 && (
-              <View style={styles.workoutStat}>
-                <Ionicons name="flame-outline" size={14} color={colors.textSecondary} />
-                <Text style={[styles.workoutStatText, { color: colors.textSecondary }]}>{post.workoutDetails.calories} cal</Text>
-              </View>
-            )}
-          </View>
-        </View>
-      )}
-      
-      {post.image && (
-        <Image 
-          source={getSafeImageUri(post.image)} 
-          style={styles.postImage} 
-          resizeMode="cover" 
-          onError={(e) => {
-            console.log('Post image error:', e.nativeEvent.error);
-            console.log('Failed image URL:', post.image);
-          }}
-        />
-      )}
-      
-      <View style={styles.postActions}>
-        <TouchableOpacity 
-          style={styles.actionButton}
-          onPress={handleLike}
-        >
-          <Ionicons 
-            name={post.isLiked ? "heart" : "heart-outline"} 
-            size={20} 
-            color={post.isLiked ? colors.primary : colors.textSecondary} 
-          />
-          <Text style={[styles.actionText, { color: colors.textSecondary }]}>{post.likes || 0} Likes</Text>
-        </TouchableOpacity>
-        <TouchableOpacity 
-          style={styles.actionButton}
-          onPress={handleComment}
-        >
-          <Ionicons name="chatbubble-outline" size={20} color={colors.textSecondary} />
-          <Text style={[styles.actionText, { color: colors.textSecondary }]}>{post.comments || 0} Comments</Text>
-        </TouchableOpacity>
-      </View>
-    </View>
-  );
-};
 
 // Simplified Progress Card - Just displays basic stats
 const ProgressCard = ({ stats, loading = false, onViewFullStats }) => {
